+++ src/ZLinq/Linq/Index.cs
@@ -50,7 +50,7 @@ namespace ZLinq.Linq
                 if (EnumeratorHelper.TryGetSlice(span, offset, destination.Length, out var slice))
                 {
                     var index = offset.GetOffset(span.Length);
-                    for (var i = 0; i < slice.Length; i++)
+                    for (var i = 0; (uint)i < (uint)slice.Length; i++)
                     {
                         destination[i] = (index, slice[i]);
                         checked { index++; }
